Access years and state properties correctly
Static methods cannot use `$this` to access properties in the class.

We can get around this by calling an instance of the class and
accessing the property directly.

This also adjusts the logic to avoid reassignment of the variable
to itself. Instead, we can assume that the value is okay unless
it fails the check.